requirements.txt
# AIsCopy - Real-time Screen Translation Tool
# Python 3.12+ required

# GUI Framework
PySide6>=6.6.0

# Screen Capture
mss>=9.0.1

# Image Processing
opencv-python>=4.12.0
Pillow>=11.3.0
numpy>=2.2.0
scikit-image>=0.25.0
scipy>=1.16.0

# Global Hotkeys
pynput>=1.7.6

# AI API
google-generativeai>=0.4.0

# Async HTTP
aiohttp>=3.9.0

# Security
cryptography>=42.0.0

# Configuration
configparser>=5.3.0
